setup: sets up fully qualified rancher server via docker-machine (see ext)
Rancher is distributed in "snapshots", which allow us to maintain strict versions for
new developers to work on.

The server is run on host, but has agents run in vms using RancherOS, which negates
issues with Docker for Mac and Docker for Windows.

Setup should work on both Linux and Mac right now, with Windows support coming.